consolidation: moved group items to common
