+++ optimus/lib/optimus.dart
@@ -31,7 +31,7 @@ export 'src/form/checkbox_form_field.dart';
 export 'src/form/date_time_form_field.dart';
 export 'src/form/input_form_field.dart';
 export 'src/form/select_form_field.dart';
-export 'src/group_item.dart';
+export 'src/common/group_item.dart';
 export 'src/icon.dart';
 export 'src/icon_list.dart';
 export 'src/input_field.dart';
