+++ optimus/lib/src/checkbox/checkbox.dart
@@ -4,7 +4,7 @@ import 'package:flutter/widgets.dart';
 import 'package:optimus/optimus.dart';
 import 'package:optimus/src/border_radius.dart';
 import 'package:optimus/src/enabled.dart';
-import 'package:optimus/src/group_wrapper.dart';
+import 'package:optimus/src/common/group_wrapper.dart';
 import 'package:optimus/src/theme/theme.dart';
 import 'package:optimus/src/typography/presets.dart';
