+++ optimus/lib/src/common/group_item.dart
