+++ optimus/lib/src/field_wrapper.dart
@@ -6,7 +6,6 @@ import 'package:optimus/src/common/field_error.dart';
 import 'package:optimus/src/common/field_label.dart';
 import 'package:optimus/src/constants.dart';
 import 'package:optimus/src/theme/theme.dart';
-import 'package:optimus/src/typography/styles.dart';
 
 class FieldWrapper extends StatefulWidget {
   const FieldWrapper({
