+++ optimus/lib/src/segmented_control/segmented_control.dart
@@ -5,7 +5,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:optimus/optimus.dart';
 import 'package:optimus/src/enabled.dart';
-import 'package:optimus/src/group_wrapper.dart';
+import 'package:optimus/src/common/group_wrapper.dart';
 import 'package:optimus/src/segmented_control/wrapper.dart';
 import 'package:optimus/src/typography/presets.dart';
