+++ optimus/lib/src/typography/styles.dart
@@ -1,79 +1,79 @@
-import 'package:flutter/rendering.dart';
-
-const TextStyle baseTextStyle = TextStyle(
-  fontFamily: 'Montserrat',
-  package: 'optimus',
-  fontWeight: FontWeight.w400,
-  height: 1.3,
-);
-
-final TextStyle preset700s = baseTextStyle.copyWith(
-  fontSize: 36,
-  fontWeight: FontWeight.w600,
-);
-
-final TextStyle preset600s = baseTextStyle.copyWith(
-  fontSize: 30,
-  fontWeight: FontWeight.w600,
-);
-
-final TextStyle preset500s = baseTextStyle.copyWith(
-  fontSize: 24,
-  fontWeight: FontWeight.w600,
-);
-
-final TextStyle preset400s = baseTextStyle.copyWith(
-  fontSize: 18,
-  fontWeight: FontWeight.w600,
-);
-
-final TextStyle preset300s = baseTextStyle.copyWith(
-  fontSize: 16,
-  height: 1.5,
-  fontWeight: FontWeight.w600,
-);
-
-final TextStyle preset300m = baseTextStyle.copyWith(
-  fontSize: 16,
-  height: 1.5,
-  fontWeight: FontWeight.w500,
-);
-
-final TextStyle preset300r = baseTextStyle.copyWith(
-  fontSize: 16,
-  height: 1.5,
-);
-
-final TextStyle preset200s = baseTextStyle.copyWith(
-  fontSize: 14,
-  height: 1.6,
-  fontWeight: FontWeight.w600,
-);
-
-final TextStyle preset200m = baseTextStyle.copyWith(
-  fontSize: 14,
-  height: 1.5,
-  fontWeight: FontWeight.w500,
-);
-
-final TextStyle preset200r = baseTextStyle.copyWith(
-  fontSize: 14,
-  height: 1.5,
-);
-
-final TextStyle preset100s = baseTextStyle.copyWith(
-  fontSize: 12,
-  height: 1.5,
-  fontWeight: FontWeight.w600,
-);
-
-final TextStyle preset100m = baseTextStyle.copyWith(
-  fontSize: 12,
-  height: 1.5,
-  fontWeight: FontWeight.w500,
-);
-
-final TextStyle preset100r = baseTextStyle.copyWith(
-  fontSize: 12,
-  height: 1.5,
-);
+// import 'package:flutter/rendering.dart';
+//
+// const TextStyle baseTextStyle = TextStyle(
+//   fontFamily: 'Montserrat',
+//   package: 'optimus',
+//   fontWeight: FontWeight.w400,
+//   height: 1.3,
+// );
+//
+// final TextStyle preset700s = baseTextStyle.copyWith(
+//   fontSize: 36,
+//   fontWeight: FontWeight.w600,
+// );
+//
+// final TextStyle preset600s = baseTextStyle.copyWith(
+//   fontSize: 30,
+//   fontWeight: FontWeight.w600,
+// );
+//
+// final TextStyle preset500s = baseTextStyle.copyWith(
+//   fontSize: 24,
+//   fontWeight: FontWeight.w600,
+// );
+//
+// final TextStyle preset400s = baseTextStyle.copyWith(
+//   fontSize: 18,
+//   fontWeight: FontWeight.w600,
+// );
+//
+// final TextStyle preset300s = baseTextStyle.copyWith(
+//   fontSize: 16,
+//   height: 1.5,
+//   fontWeight: FontWeight.w600,
+// );
+//
+// final TextStyle preset300m = baseTextStyle.copyWith(
+//   fontSize: 16,
+//   height: 1.5,
+//   fontWeight: FontWeight.w500,
+// );
+//
+// final TextStyle preset300r = baseTextStyle.copyWith(
+//   fontSize: 16,
+//   height: 1.5,
+// );
+//
+// final TextStyle preset200s = baseTextStyle.copyWith(
+//   fontSize: 14,
+//   height: 1.6,
+//   fontWeight: FontWeight.w600,
+// );
+//
+// final TextStyle preset200m = baseTextStyle.copyWith(
+//   fontSize: 14,
+//   height: 1.5,
+//   fontWeight: FontWeight.w500,
+// );
+//
+// final TextStyle preset200r = baseTextStyle.copyWith(
+//   fontSize: 14,
+//   height: 1.5,
+// );
+//
+// final TextStyle preset100s = baseTextStyle.copyWith(
+//   fontSize: 12,
+//   height: 1.5,
+//   fontWeight: FontWeight.w600,
+// );
+//
+// final TextStyle preset100m = baseTextStyle.copyWith(
+//   fontSize: 12,
+//   height: 1.5,
+//   fontWeight: FontWeight.w500,
+// );
+//
+// final TextStyle preset100r = baseTextStyle.copyWith(
+//   fontSize: 12,
+//   height: 1.5,
+// );
